add profile com
