+++ app/components/MainLayout.tsx
@@ -149,10 +149,10 @@ const MainLayout = ({children}:{
 
         <li className='flex justify-start items-center  text-[#ecd6fb] hover:bg-[#ecd6fb] hover:text-[#33265e] rounded-xl p-2'>
 <BsInfoSquare className='mr-2 [#ecd6fb]'/>
-<Link href="../aboutibd"  onClick={closeMenue}>⁠About IBD</Link>
+<Link href="../pages/aboutibd"  onClick={closeMenue}>⁠About IBD</Link>
         </li> <li className='flex justify-start items-center  text-[#ecd6fb] hover:bg-[#ecd6fb] hover:text-[#33265e] rounded-xl p-2'>
 <MdOutlineContactSupport className='mr-2 [#ecd6fb]'/>
-<Link href="../aihelp" onClick={closeMenue}>⁠AI Help</Link>
+<Link href="../pages/aihelp" onClick={closeMenue}>⁠AI Help</Link>
         </li>   
     
         <li className='flex flex-col justify-start items-start rounded-xl p-2'>
@@ -167,11 +167,11 @@ const MainLayout = ({children}:{
 <ul className='flex flex-col justify-start items-start ml-4 mt-4 '>
     <li className='flex justify-center items-center  text-[#ecd6fb] gap-4 hover:bg-[#ecd6fb] hover:text-[#33265e] rounded-xl p-2 '>
 <FaUserDoctor className='mr-2  [#ecd6fb]'/>
-<Link href={"../doctors"} onClick={closeMenue}>Doctors</Link>
+<Link href={"../pages/doctors"} onClick={closeMenue}>Doctors</Link>
     </li>
     <li className='flex justify-center items-center gap-4  text-[#ecd6fb] hover:bg-[#ecd6fb] hover:text-[#33265e] rounded-xl p-2'>
 <LuApple className='mr-2  [#ecd6fb]'/>
-<Link href={"../dietitian"} onClick={closeMenue}>⁠Dietitian</Link>
+<Link href={"../pages/dietitian"} onClick={closeMenue}>⁠Dietitian</Link>
     </li>
     <li className='flex justify-center items-center gap-4 text-[#ecd6fb] hover:bg-[#ecd6fb] hover:text-[#33265e] rounded-xl p-2'>
 <IoNavigateCircleOutline className='mr-2  [#ecd6fb]'/>
@@ -179,15 +179,15 @@ const MainLayout = ({children}:{
     </li>
     <li className='flex justify-center items-center gap-4  text-[#ecd6fb] hover:bg-[#ecd6fb] hover:text-[#33265e] rounded-xl p-2'>
 <MdOutlineManageAccounts className='mr-2  [#ecd6fb]'/>
-<Link href={"../manageibd"} onClick={closeMenue}>⁠Manage IBD</Link>
+<Link href={"../pages/manageibd"} onClick={closeMenue}>⁠Manage IBD</Link>
     </li>
     <li className='flex justify-center items-center gap-4  text-[#ecd6fb] hover:bg-[#ecd6fb] hover:text-[#33265e] blue-800 rounded-xl p-2'>
 <MdOutlineArticle className='mr-2  [#ecd6fb]'/>
-<Link href={"../articles"} onClick={closeMenue}>⁠Articles</Link>
+<Link href={"../pages/articles"} onClick={closeMenue}>⁠Articles</Link>
     </li>
     <li className='flex justify-center items-center gap-4  text-[#ecd6fb] hover:bg-[#ecd6fb] hover:text-[#33265e] rounded-xl p-2'>
 <GoOrganization className='mr-2  [#ecd6fb]'/>
-<Link href={"../organizations"} onClick={closeMenue}>⁠Organizations</Link>
+<Link href={"../pages/organizations"} onClick={closeMenue}>⁠Organizations</Link>
     </li>
 </ul>
 }
@@ -197,11 +197,11 @@ const MainLayout = ({children}:{
 
         <li className='flex justify-start items-center  text-[#ecd6fb] hover:bg-[#ecd6fb] hover:text-[#33265e] rounded-xl p-2'>
 <SiHelpscout className='mr-2  [#ecd6fb]'/>
-<Link href="../aboutus" onClick={closeMenue}>About us</Link>
+<Link href="../pages/aboutus" onClick={closeMenue}>About us</Link>
         </li>
         <li className='flex justify-start items-center  text-[#ecd6fb] hover:bg-[#ecd6fb] hover:text-[#33265e] rounded-xl p-2'>
 <FiPhoneCall className='mr-2  [#ecd6fb]'/>
-<Link href="../contact-us" onClick={closeMenue}>Contact us</Link>
+<Link href="../pages/contact-us" onClick={closeMenue}>Contact us</Link>
         </li>
 
     </ul>
