--- app/doctors/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>Doctors</div>
-  )
-}
-
-export default page
