--- app/donate/page.tsx
@@ -1,20 +1,0 @@
-import React from 'react'
-import Image from "next/image";
-
-//import { Spotlight } from '../ui/Spotlight'
-
-const page = () => {
-  return (
-    <div className="flex flex-col w-full" >
-      
-      
-   Donate
-    </div>
-   
-  )
-}
-
-export default page
-
-
-////////////
