--- app/manageibd/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>Manage IBD</div>
-  )
-}
-
-export default page
