--- app/organizations/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>Organization</div>
-  )
-}
-
-export default page
