--- app/toiletnavigator/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>⁠Toiletnavigator</div>
-  )
-}
-
-export default page
