+++ app/ui/typewriter-effect.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from"../components/lib/utils";
+import { cn } from"../../components/lib/utils";
 import { motion, stagger, useAnimate, useInView } from "framer-motion";
 import { useEffect } from "react";
